requirements.txt
Django>=5.0
django-unfold>=0.24.0
django-celery-beat>=2.6.0
celery[redis]>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
django-debug-toolbar>=4.3.0
logtail-python>=0.2.0